corrections

+++ The-Tower/Assets/Scripts/Fase0.cs
@@ -14,11 +14,11 @@ public class Fase0 : MonoBehaviour {
 		
 	}
     public void StartGame() {
-        PlayerPrefs.SetInt("Hp", 20);
-        PlayerPrefs.SetInt("Dex", 5);
-        PlayerPrefs.SetInt("Str", 5);
-        PlayerPrefs.SetInt("Def", 0);
-        PlayerPrefs.SetInt("Aura", 0);
+        PlayerPrefs.SetFloat("Hp", 20);
+        PlayerPrefs.SetFloat("Dex", 5);
+        PlayerPrefs.SetFloat("Str", 5);
+        PlayerPrefs.SetFloat("Def", 0);
+        PlayerPrefs.SetFloat("Aura", 0);
         PlayerPrefs.SetInt("Andar", 1);
 
         for (int i = 0; i < 8; i++)
